+++ Assets/Mu3LibraryAssets/Samples~/Sample_Template/Scripts/Runtime/Sample_Localization/LocalizationTableKeys.cs
@@ -1,0 +1,16 @@
+﻿namespace Mu3Library.Sample.Template.Localization
+{
+    public static class LocalizationTableKeys
+    {
+        public static class TestStringTable
+        {
+            public const string Name = "TestStringTable";
+            
+            public static class hello
+            {
+                public const string Key = "hello";
+                public const string Id = "909386731520";
+            }
+        }
+    }
+}